Show how to achieve HMR support for babel-plugin-css-in-js
This code that made it work are the lines added to src/index.js. This is kind of a hack but since hot module reloading is only used in development I can certainly live with that.

It may be a good idea to integrate an improved version of this as a shipped module of the plugin but I don't have the time at the moment.